# server/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "HTTP API for Phase 1 user top-up campaigns (admin + user-facing). Operates under `/campaign-center-api/v1`;",
        "title": "Campaign Center API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/campaign-center-api/v1",
    "paths": {
...
        "/admin/campaigns/{campaignId}/participations": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-campaign-performance"
                ],
                "summary": "List campaign participations (admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Campaign ID",
                        "name": "campaignId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size (default 20)",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by user ID",
                        "name": "userId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by reward status e.g. GRANTED",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/data.StandardResponse"
                        }
                    },
                    "404": {
                        "description": "campaign not found",
                        "schema": {
                            "$ref": "#/definitions/data.StandardResponse"
                        }
                    },
                    "503": {
                        "description": "database unavailable",
                        "schema": {
                            "$ref": "#/definitions/data.StandardResponse"
                        }
                    }
                }
            }
        },
        "/admin/campaigns/{campaignId}/performance/daily": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-campaign-performance"
                ],
                "summary": "List campaign daily performance (admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Campaign ID",
                        "name": "campaignId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Start date YYYY-MM-DD",
                        "name": "startDate",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End date YYYY-MM-DD",
                        "name": "endDate",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/data.StandardResponse"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "$ref": "#/definitions/data.StandardResponse"
                        }
                    },
                    "404": {
                        "description": "campaign not found",
                        "schema": {
                            "$ref": "#/definitions/data.StandardResponse"
                        }
                    },
                    "503": {
                        "description": "database unavailable",
                        "schema": {
                            "$ref": "#/definitions/data.StandardResponse"
                        }
                    }
                }
            }
        },
        "/admin/campaigns/{campaignId}/performance/summary": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-campaign-performance"
                ],
                "summary": "Get campaign performance summary (admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Campaign ID",
                        "name": "campaignId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/data.StandardResponse"
                        }
                    },
                    "404": {
                        "description": "campaign not found",
                        "schema": {
                            "$ref": "#/definitions/data.StandardResponse"
                        }
                    },
                    "503": {
                        "description": "database unavailable",
                        "schema": {
                            "$ref": "#/definitions/data.StandardResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "data.StandardResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "success"
                }
            }
        }
    }
}